+++ src/module/NukeXParameters.cpp
@@ -129,7 +129,7 @@ NXBackgroundTarget::NXBackgroundTarget( MetaProcess* p ) : MetaFloat( p )
 
 IsoString NXBackgroundTarget::Id() const { return "backgroundTarget"; }
 int       NXBackgroundTarget::Precision() const { return 3; }
-double    NXBackgroundTarget::DefaultValue() const { return 0.25; }
+double    NXBackgroundTarget::DefaultValue() const { return 0.12; }
 double    NXBackgroundTarget::MinimumValue() const { return 0.05; }
 double    NXBackgroundTarget::MaximumValue() const { return 0.50; }